Rename types.py to models.py
