+++ integration/test_rbac.py
@@ -3,7 +3,7 @@ from weaviate.auth import Auth
 from weaviate.rbac.types import RBAC
 
 
-def test_rbac(client_factory: ClientFactory) -> None:
+def test_create_role(client_factory: ClientFactory) -> None:
     with client_factory(
         ports=(8080, 50051), auth_credentials=Auth.api_key("jp-secret-key")
     ) as client:
@@ -11,8 +11,8 @@ def test_rbac(client_factory: ClientFactory) -> None:
             name="CollectionCreator",
             permissions=RBAC.permissions.database(actions=RBAC.actions.database.CREATE_COLLECTION),
         )
-        roles = client.roles.list_all()
-        assert len(roles) == 1
-        assert roles[0].name == "CollectionCreator"
-        assert roles[0].database_permissions is not None
-        assert roles[0].database_permissions[0].actions == [RBAC.actions.database.CREATE_COLLECTION]
+        role = client.roles.by_name("CollectionCreator")
+        assert role is not None
+        assert role.name == "CollectionCreator"
+        assert role.database_permissions is not None
+        assert role.database_permissions[0].actions == [RBAC.actions.database.CREATE_COLLECTION]
